tunit_t is now uint64_t
+ Largely inconsequential, but might make more sense for 32 bit
  platforms

+++ include/kmi/timer.h
@@ -25,7 +25,7 @@ typedef uint64_t ticks_t;
 /**
  * tunit_t typedef, whichever time unit we're dealing with.
  */
-typedef size_t tunit_t;
+typedef uint64_t tunit_t;
 
 /**
  * Timer structure.
